Gestione risposte http del PostController
- Gestione risposte http del PostControllerutilizzando ResponseEntity
- deletePostById non restituisce un body nel caso vada a buon fine (204 No Content)

+++ Api-MicroBlog/src/main/java/com/bachir/apimicroblog/controller/PostController.java
@@ -6,12 +6,16 @@
 package com.bachir.apimicroblog.controller;
 
 import com.bachir.apimicroblog.domain.Post;
+import com.bachir.apimicroblog.entities.JsonResponseBody;
 import com.bachir.apimicroblog.service.PostService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import java.util.List;
+import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -30,32 +34,73 @@ public class PostController
     @Autowired
     PostService postService;
     
-    @RequestMapping(method = RequestMethod.GET)
+    @RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation(value = "returns all the posts")
-    public List<Post> getPosts()
+    public ResponseEntity<JsonResponseBody> getPosts()
     {
-        return postService.getAllPosts();
+        try
+        {
+            List<Post> posts = postService.getAllPosts();
+            return ResponseEntity.status(HttpStatus.OK).body(new JsonResponseBody(HttpStatus.OK.value(), posts));
+        }
+        catch( Exception e )
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new JsonResponseBody(HttpStatus.BAD_REQUEST.value(), "Error: " + e.toString()));
+        }
     }
     
     @RequestMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.GET)
     @ApiOperation(value = "returns a post by an ID")
-    public Post getPostById(@ApiParam@PathVariable("id") Long postId)
+    public ResponseEntity<JsonResponseBody> getPostById(@ApiParam@PathVariable("id") Long postId)
     {
-        return postService.getPostById(postId);
+        try
+        {
+            
+            Post post = postService.getPostById(postId).get();
+            return ResponseEntity.status(HttpStatus.FOUND).body(new JsonResponseBody(HttpStatus.FOUND.value(), post));
+        }
+        catch( Exception e )
+        {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new JsonResponseBody(HttpStatus.NOT_FOUND.value(), "Error: " + e.toString()));
+        }
     }
     
     @RequestMapping(method = RequestMethod.POST)
-    @ApiOperation(value = "adds a post in the database")
-    public void addPost(@ApiParam@RequestBody Post post)
+    @ApiOperation(value = "adds a post in the database", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<JsonResponseBody> addPost(HttpServletRequest request, @ApiParam@RequestBody Post p)
     {
-        postService.insertPost(post);
+        try
+        {   
+            Post post = postService.insertPost(p);
+            return ResponseEntity.status(HttpStatus.CREATED).header("location", request.getRequestURL().toString() + post.getId()).body(new JsonResponseBody(HttpStatus.CREATED.value(), null));
+        }
+        catch( Exception e )
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new JsonResponseBody(HttpStatus.BAD_REQUEST.value(), "Error: " + e.toString()));
+        }
     }
     
-    @RequestMapping(method = RequestMethod.DELETE,value = "/deletepost/{id}")
+    @RequestMapping(method = RequestMethod.DELETE,value = "/deletepost/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation(value = "deletes a post by an ID")
-    public void deletePostById(@ApiParam@PathVariable("id") Long postId)
+    public ResponseEntity<JsonResponseBody> deletePostById(@ApiParam@PathVariable("id") Long postId)
     {
-        postService.deletePostById(postId);
+        try
+        {   
+            if( postService.getPostById(postId).isPresent())
+            {
+                postService.deletePostById(postId);
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new JsonResponseBody(HttpStatus.NO_CONTENT.value(), null));
+            }
+            else
+            {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new JsonResponseBody(HttpStatus.NOT_FOUND.value(), "Error: post not found"));
+            }
+                
+        }
+        catch( Exception e )
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new JsonResponseBody(HttpStatus.BAD_REQUEST.value(), "Error: " + e.toString()));
+        }
     }
     
 }
